+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.4.0 - 2025-03-11
+- Add option to hide Minecraft window
+- Add clear_input action
+- Add raw frame support and make it the default
+- Remove png and jpeg frame types
+- Support other MCio 0.4.0 changes
+- MCIO_PROTOCOL_VERSION = 3
+
 ## 0.3.2 - 2025-01-15
 - Silence log warning during install
 - MCIO_PROTOCOL_VERSION = 2